+++ cspatients/tests/test_views.py
@@ -1,9 +1,6 @@
 from os import environ
 
-import pytest
 import responses
-from channels.layers import get_channel_layer
-from channels.testing import WebsocketCommunicator
 from django.conf import settings
 from django.contrib.auth.models import User
 from django.test import Client, TestCase
@@ -16,160 +13,21 @@ from rest_framework import status
 from rest_framework.authtoken.models import Token
 from rest_framework.test import APIClient, APITestCase
 
-from .consumers import ViewConsumer
-from .models import Baby, PatientEntry, Profile
-from .util import (
-    get_all_active_patient_entries,
-    get_rp_dict,
-    save_model,
-    save_model_changes,
+from cspatients.models import Baby, PatientEntry, Profile
+
+from .constants import (
+    SAMPLE_RP_CHECKLIST_DATA,
+    SAMPLE_RP_CHECKLIST_DATA_INACTIVE,
+    SAMPLE_RP_POST_DATA,
+    SAMPLE_RP_POST_DATA_NON_EXISTING,
+    SAMPLE_RP_POST_INVALID_DATA,
+    SAMPLE_RP_POST_NO_CONSENT_DATA,
+    SAMPLE_RP_UPDATE_COMPLETED_DATA,
+    SAMPLE_RP_UPDATE_DATA,
+    SAMPLE_RP_UPDATE_DELIVERY_DATA,
+    SAMPLE_RP_UPDATE_DELIVERY_DATA_MINIMAL,
+    SAMPLE_RP_UPDATE_INVALID_DATA,
 )
-
-# View Tests
-
-SAMPLE_RP_POST_DATA = {
-    "results": {
-        "surname": {"category": "All Responses", "value": "Doe", "input": "Doe"},
-        "decision_time": {
-            "category": "All Responses",
-            "value": "2019-05-12 10:22+00:00",
-            "input": "2019-05-12 10:22+00:00",
-        },
-        "option": {"category": "Patient Entry", "value": "1", "input": "1"},
-    }
-}
-
-SAMPLE_RP_POST_NO_CONSENT_DATA = {
-    "results": {
-        "surname": {
-            "category": "All Responses",
-            "value": "(No Consent)",
-            "input": "(No Consent)",
-        },
-        "gravidity": {"category": "All Responses", "value": "1", "input": "1"},
-        "parity": {"category": "All Responses", "value": "2", "input": "2"},
-        "option": {"category": "Patient Entry", "value": "1", "input": "1"},
-        "clinician": {
-            "category": "All Responses",
-            "value": "Dr Test",
-            "input": "Dr Test",
-        },
-        "consent": {
-            "name": "Consent",
-            "category": "no_consent",
-            "value": "2",
-            "input": "2",
-        },
-    }
-}
-
-SAMPLE_RP_UPDATE_DATA = {
-    "results": {
-        "patient_id": {"category": "All Responses", "value": "1", "input": "1"},
-        "new_value": {
-            "category": "All Responses",
-            "value": "Nyasha",
-            "input": "Nyasha",
-        },
-        "change_category": {"category": "surname", "value": "1", "input": "1"},
-        "option": {"category": "Patient Entry", "value": "1", "input": "1"},
-    }
-}
-
-SAMPLE_RP_POST_DATA_NON_EXISTING = {
-    "results": {
-        "name": {"category": "All Responses", "value": "Jane Moe", "input": "Jane Moe"},
-        "patient_id": {"category": "All Responses", "value": "999", "input": "999"},
-        "change_category": {"category": "name", "value": "A", "input": "A"},
-        "new_value": {
-            "category": "All Responses",
-            "value": "Nyasha",
-            "input": "Nyasha",
-        },
-    }
-}
-
-SAMPLE_RP_POST_INVALID_DATA = {
-    "results": {
-        "name": {"category": "All Responses", "value": "Jane Doe", "input": "Jane Moe"},
-        "gravpar_invalid": {
-            "category": "All Responses",
-            "value": "abcdef",
-            "input": "abcdef",
-        },
-    }
-}
-
-SAMPLE_RP_UPDATE_INVALID_DATA = {
-    "results": {
-        "new_value": {
-            "category": "All Responses",
-            "value": "Nyasha",
-            "input": "Nyasha",
-        },
-        "change_category": {"category": "surname", "value": "1", "input": "1"},
-        "option": {"category": "Patient Entry", "value": "1", "input": "1"},
-    }
-}
-
-SAMPLE_RP_UPDATE_DELIVERY_DATA = {
-    "results": {
-        "patient_id": {"category": "All Responses", "value": "1", "input": "1"},
-        "foetus": {"name": "foetus", "category": "All Responses", "value": "2"},
-        "baby_number": {
-            "name": "baby_number",
-            "category": "All Responses",
-            "value": "2",
-        },
-        "delivery_time": {
-            "name": "delivery_time",
-            "category": "All Responses",
-            "value": "2019-05-12 10:22+00:00",
-        },
-        "baby_weight_grams": {
-            "name": "baby_weight_grams",
-            "category": "All Responses",
-            "value": "2400",
-        },
-        "nicu": {"name": "nicu", "category": "Yes", "value": "2"},
-        "apgar_1": {"name": "apgar_1", "category": "All Responses", "value": "8"},
-        "apgar_5": {"name": "apgar_5", "category": "All Responses", "value": "9"},
-        "option": {"category": "Delivery", "value": "3", "input": "3"},
-    }
-}
-
-SAMPLE_RP_UPDATE_DELIVERY_DATA_MINIMAL = {
-    "results": {
-        "patient_id": {"category": "All Responses", "value": "HLFSH", "input": "HLFSH"},
-        "foetus": {"name": "foetus", "category": "All Responses", "value": "2"},
-        "baby_number": {
-            "name": "baby_number",
-            "category": "All Responses",
-            "value": "1",
-        },
-        "delivery_time": {
-            "name": "delivery_time",
-            "category": "All Responses",
-            "value": "2019-05-12 10:22+00:00",
-        },
-        "option": {"category": "Delivery", "value": "3", "input": "3"},
-    }
-}
-
-SAMPLE_RP_UPDATE_COMPLETED_DATA = {
-    "results": {
-        "patient_id": {"category": "All Responses", "value": "HLFSH", "input": "HLFSH"},
-        "completion_time": {
-            "name": "completion_time",
-            "category": "All Responses",
-            "value": "2019-05-12 10:22+00:00",
-        },
-        "option": {"category": "Completed", "value": "4", "input": "4"},
-    }
-}
-
-SAMPLE_RP_CHECKLIST_DATA = {"contact": {"urn": "whatsapp:12065550109"}}
-SAMPLE_RP_CHECKLIST_DATA_INACTIVE = {"contact": {"urn": "whatsapp:12065550108"}}
 
 
 class FormTest(TestCase):
@@ -313,233 +171,6 @@ class PatientViewTest(TestCase):
             response=response, template_name="cspatients/patient.html"
         )
         self.assertInHTML("Jane", str(response.content))
-
-
-# Util Tests
-
-
-class ViewAllContextTest(TestCase):
-    def setUp(self):
-        self.patient_one_data = {"surname": "Urgency COLD", "age": 20}
-
-        self.patient_two_data = {
-            "surname": "Urgency Immediate",
-            "age": 23,
-            "urgency": 1,
-        }
-
-        self.patient_three_data = {"surname": "John Completed", "age": 20}
-
-        self.patient_four_data = {
-            "surname": "Urgency Immediate NEW",
-            "age": 23,
-            "urgency": 1,
-        }
-
-        self.patient_five_data = {"surname": "Completed OLD", "age": 23, "urgency": 1}
-
-    def test_context_when_no_patients(self):
-        self.assertFalse(get_all_active_patient_entries())
-
-    @freeze_time("08:00")
-    def test_get_all_active_patient_entries(self):
-        save_model(self.patient_one_data)
-        save_model(self.patient_two_data)
-        entry3, _ = save_model(self.patient_three_data)
-        entry4, _ = save_model(self.patient_four_data)
-        entry5, _ = save_model(self.patient_five_data)
-
-        entry3.completion_time = timezone.now()
-        entry3.save()
-
-        entry4.decision_time = timezone.now() + timezone.timedelta(hours=1)
-        entry4.save()
-
-        entry5.completion_time = timezone.now()
-        entry5.decision_time = timezone.now() - timezone.timedelta(days=1)
-        entry5.save()
-
-        patient_entries = get_all_active_patient_entries()
-
-        # Check that it returns 4 objects
-        self.assertEqual(len(patient_entries), 4)
-
-        # Check that the results are sorted by the urgency
-        self.assertEqual(patient_entries[0].surname, "Urgency Immediate NEW")
-        self.assertEqual(patient_entries[1].surname, "Urgency Immediate")
-        self.assertEqual(patient_entries[2].surname, "Urgency COLD")
-        self.assertEqual(patient_entries[3].surname, "John Completed")
-
-    @freeze_time("04:30")
-    def test_get_all_active_patient_entries_before_5_utc(self):
-        save_model(self.patient_one_data)
-        save_model(self.patient_two_data)
-        entry3, _ = save_model(self.patient_three_data)
-        entry4, _ = save_model(self.patient_four_data)
-        entry5, _ = save_model(self.patient_five_data)
-
-        entry3.completion_time = timezone.now()
-        entry3.save()
-
-        entry4.decision_time = timezone.now() + timezone.timedelta(hours=1)
-        entry4.save()
-
-        entry5.completion_time = timezone.now()
-        entry5.decision_time = timezone.now() - timezone.timedelta(days=1)
-        entry5.save()
-
-        patient_entries = get_all_active_patient_entries()
-
-        # Check that it returns 4 objects
-        self.assertEqual(len(patient_entries), 5)
-
-        # Check that the results are sorted by the urgency
-        self.assertEqual(patient_entries[0].surname, "Urgency Immediate NEW")
-        self.assertEqual(patient_entries[1].surname, "Urgency Immediate")
-        self.assertEqual(patient_entries[2].surname, "Urgency COLD")
-        self.assertEqual(patient_entries[3].surname, "Completed OLD")
-        self.assertEqual(patient_entries[4].surname, "John Completed")
-
-    def test_get_all_active_patient_entries_filter(self):
-        save_model(self.patient_one_data)
-        save_model(self.patient_two_data)
-        save_model(self.patient_three_data)
-
-        patient_entries = get_all_active_patient_entries("COLD", "4")
-
-        self.assertEqual(len(patient_entries), 1)
-        self.assertEqual(patient_entries[0].surname, "Urgency COLD")
-
-    def test_get_all_active_patient_entries_filter_complete(self):
-        save_model(self.patient_one_data)
-        save_model(self.patient_two_data)
-        entry3, _ = save_model(self.patient_three_data)
-
-        entry3.completion_time = timezone.now()
-        entry3.save()
-
-        patient_entries = get_all_active_patient_entries(None, "complete")
-
-        self.assertEqual(len(patient_entries), 1)
-        self.assertEqual(patient_entries[0].surname, "John Completed")
-
-
-class SaveModelTest(TestCase):
-    def setUp(self):
-        self.patient_one_data = {"surname": "Jane Doe", "age": 20, "urgency": 1}
-
-    def test_saves_when_given_sufficient_data(self):
-        patiententry, errors = save_model(self.patient_one_data)
-
-        # Must return the saved patient entry
-        self.assertEqual(errors, [])
-        self.assertIsNotNone(patiententry)
-
-        # Check that the save persisted in the database
-        self.assertEqual(len(PatientEntry.objects.all()), 1)
-
-    def test_saves_when_given_junk_data(self):
-        patiententry, errors = save_model(
-            {"surname": "Mary Mary", "age": 23, "urgency": 1, "response_19": "test"}
-        )
-
-        # Must return the saved patient entry
-        self.assertIsNotNone(patiententry)
-        self.assertEqual(errors, [])
-
-        # Check that the save persisted in the database
-        self.assertEqual(len(PatientEntry.objects.all()), 1)
-
-    def test_nothing_saved_with_insufficient_data(self):
-        patientryentry, errors = save_model({"age": 23, "urgency": 1})
-
-        # Must return None
-        self.assertIsNone(patientryentry)
-        self.assertEqual(errors, ["Surname is required"])
-
-        # Check that none of the information is in the database
-        self.assertEqual(len(PatientEntry.objects.all()), 0)
-
-
-class SaveModelChangesTest(TestCase):
-    def setUp(self):
-        self.patient_entry = PatientEntry.objects.create(surname="Doe", age=20)
-
-    def test_saves_data_when_passed_good_dict(self):
-        changes_dict = {
-            "patient_id": str(self.patient_entry.id),
-            "surname": "Moe",
-            "gravidity": "1",
-            "parity": "1",
-        }
-
-        # Check returns PatientEntry
-        result, errors = save_model_changes(changes_dict)
-        self.assertEqual(errors, [])
-        self.assertTrue(isinstance(result, PatientEntry))
-
-        # Check that the changes persisted.
-        self.patient_entry.refresh_from_db()
-        self.assertEqual(self.patient_entry.surname, "Moe")
-        self.assertEqual(self.patient_entry.gravpar, "G1P1")
-        self.assertEqual(str(self.patient_entry), "Moe having CS")
-
-    def test_returns_none_for_patient_dict_with_wrong_patient_id(self):
-        changes_dict = {"patient_id": "999", "surname": "Moe"}
-
-        result, errors = save_model_changes(changes_dict)
-
-        # Returns none on bad incorrect patient_id
-        self.assertIsNone(result)
-        self.assertEqual(errors, ["Patient entry does not exist"])
-
-    def test_does_not_make_changes_when_dict_has_wrong_fields(self):
-        changes_dict = {
-            "patient_id": str(self.patient_entry.id),
-            "surname": "Moe",
-            "urgent": "DSHLSD",
-        }
-
-        result, errors = save_model_changes(changes_dict)
-
-        self.assertEqual(errors, [])
-
-        self.patient_entry.refresh_from_db()
-        self.assertEqual(self.patient_entry.urgency, result.urgency)
-
-
-class GetRPDictTest(TestCase):
-
-    """
-        Test the get_rp_dict method
-        Takes in request.POST method
-    """
-
-    def test_get_rp_event_dict(self):
-        """
-            The Method should be able to extract the surname, and patient ID,
-            from the Query dict into a dictionary with just name and
-            patient_id
-        """
-        SAMPLE_RP_POST_DATA["results"]["patient_id"] = {
-            "category": "All Responses",
-            "value": "123",
-            "input": "123",
-        }
-        event_dict = get_rp_dict(SAMPLE_RP_POST_DATA)
-
-        self.assertEqual(event_dict["surname"], "Doe")
-        self.assertEqual(event_dict["patient_id"], "123")
-
-    def test_get_entrychanges_dict(self):
-        """
-            The Method when passed a context of entrychanges, must be able
-            to pick up the change category and the new value
-        """
-        changes_dict = get_rp_dict(SAMPLE_RP_UPDATE_DATA, context="entrychanges")
-
-        self.assertEqual(changes_dict["surname"], "Nyasha")
-        self.assertEqual(changes_dict["patient_id"], "1")
 
 
 # API Tests
@@ -1078,31 +709,3 @@ class DetailedHealthViewTest(APITestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json()["update"], "queues not checked")
         self.assertEqual(response.json()["queues"], [])
-
-
-class TestModels(TestCase):
-    def test_profile(self):
-        new_user = User.objects.create_user(
-            "test_user", "testuser@example.com", "1234", is_active=False
-        )
-        profile = Profile.objects.create(**{"user": new_user, "msisdn": "+12065550109"})
-
-        self.assertEqual(str(profile), "test_user: +12065550109")
-
-
-@pytest.mark.asyncio
-async def test_view_consumer():
-    communicator = WebsocketCommunicator(ViewConsumer, "/ws/cspatients/viewsocket/")
-
-    connected, subprotocol = await communicator.connect()
-    assert connected
-
-    channel_layer = get_channel_layer()
-    await channel_layer.group_send(
-        "view", {"type": "view.update", "content": "Testing"}
-    )
-
-    response = await communicator.receive_from()
-    assert response == "Testing"
-
-    await communicator.disconnect()
